+++ include/cli/detail/inputdevice.h
@@ -27,32 +27,46 @@
  * DEALINGS IN THE SOFTWARE.
  ******************************************************************************/
 
-#ifndef CLI_KEYBOARD_H_
-#define CLI_KEYBOARD_H_
+#ifndef CLI_DETAIL_INPUTDEVICE_H_
+#define CLI_DETAIL_INPUTDEVICE_H_
 
-#if defined(__unix__) || defined(__unix) || defined(__linux__)
-    #define OS_LINUX
-#elif defined(WIN32) || defined(_WIN32) || defined(_WIN64)
-    #define OS_WIN
-#elif defined(__APPLE__) || defined(__MACH__)
-    #define OS_MAC
-#else
-    #error "Platform not supported (yet)."
-#endif
+#include <functional>
+#include <string>
+#include "boostasio.h"
 
-#if defined(OS_LINUX) || defined(OS_MAC)
-    #include "linuxkeyboard.h"
-    namespace cli { using Keyboard = LinuxKeyboard; }
-#elif defined(OS_WIN)
-    #include "winkeyboard.h"
-    namespace cli { using Keyboard = WinKeyboard; }
-#else
-    #error "Platform not supported (yet)."
-#endif
+namespace cli
+{
+namespace detail
+{
 
-#undef OS_LINUX
-#undef OS_WIN
-#undef OS_MAC
+enum class KeyType { ascii, up, down, left, right, backspace, canc, home, end, ret, eof, ignored };
 
-#endif // CLI_KEYBOARD_H_
+class InputDevice
+{
+public:
+    using Handler = std::function< void( std::pair<KeyType,char> ) >;
+
+    InputDevice(detail::asio::BoostExecutor ex) : executor(ex) {}
+    virtual ~InputDevice() = default;
+
+    template <typename H>
+    void Register(H&& h) { handler = std::forward<H>(h); }
+
+protected:
+
+    void Notify(std::pair<KeyType,char> k)
+    {
+        executor.Post([this,k](){ if (handler) handler(k); });
+    }
+
+private:
+
+    detail::asio::BoostExecutor executor;
+    Handler handler;
+};
+
+} // namespace detail
+} // namespace cli
+
+#endif // CLI_DETAIL_INPUTDEVICE_H_