+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Current Version: 1.1.1-dirty
  - Fix Backspace from remote terminal (issue #52)
  - Escaping and sentence support with quote and double quote
  - Fix duplicate autocompletion (issue #67)
+ - Add a namespace and folder "detail" (issue #17)
  
 ## [1.1.1] - 2019-09-16